+++ src/config.mjs
@@ -1,0 +1,13 @@
+// @flow
+
+let cssRootFolder: string = '/public/styles/'
+
+export default class Config {
+  static get cssRootFolder() {
+    return cssRootFolder
+  }
+
+  static set cssRootFolder(dir: string) {
+    cssRootFolder = dir
+  }
+}