+++ src/index.mjs
@@ -2,6 +2,7 @@
 
 import I18n from './intl/i18n.mjs'
 import Router from './router/router.mjs'
+import Config from './config.mjs'
 import Component from './component/component.mjs'
 import { html } from './template/template.mjs'
 import { createState } from './state/state.mjs'
@@ -11,6 +12,7 @@ export {
   html,
   I18n,
   Router,
+  Config,
   Component,
   registerCss,
   createState,